refactor: register to signup
