+++ MyFinance-Backend/MyFinance.Application/UseCases/Users/Commands/SignUp/SignUpHandler.cs
@@ -2,23 +2,24 @@
 using MyFinance.Application.Abstractions.Persistence.Repositories;
 using MyFinance.Application.Abstractions.RequestHandling.Commands;
 using MyFinance.Application.Abstractions.Services;
+using MyFinance.Application.UseCases.Users.Commands.SignUp;
 using MyFinance.Domain.Entities;
 
-namespace MyFinance.Application.UseCases.Users.Commands.RegisterUser;
+namespace MyFinance.Application.UseCases.Users.Commands.SignUp;
 
-internal sealed class RegisterUserHandler(
+internal sealed class SignUpHandler(
     ITokenProvider tokenProvider,
     IUserRepository userRepository,
     IPasswordManager passwordManager,
     IEmailSender emailSender)
-    : ICommandHandler<RegisterUserCommand>
+    : ICommandHandler<SignUpCommand>
 {
     private readonly ITokenProvider _tokenProvider = tokenProvider;
     private readonly IPasswordManager _passwordManager = passwordManager;
     private readonly IUserRepository _userRepository = userRepository;
     private readonly IEmailSender _emailSender = emailSender;
 
-    public async Task<Result> Handle(RegisterUserCommand command, CancellationToken cancellationToken)
+    public async Task<Result> Handle(SignUpCommand command, CancellationToken cancellationToken)
     {
         var passwordHash = _passwordManager.HashPassword(command.PlainTextPassword);
         var user = new User(command.Name, command.Email, passwordHash);
