+++ MyFinance-Backend/MyFinance.Contracts/User/Requests/SignUpRequest.cs
@@ -1,6 +1,6 @@
 ﻿namespace MyFinance.Contracts.User.Requests;
 
-public sealed record RegisterUserRequest(
+public sealed record SignUpRequest(
     string Name,
     string Email,
     string PlainTextPassword,
